version type just uses b4
