+++ primitives/bytes/b4.go
@@ -22,6 +22,7 @@
 package bytes
 
 import (
+	"encoding/binary"
 	"fmt"
 
 	"github.com/berachain/beacon-kit/primitives/encoding/hex"
@@ -48,6 +49,18 @@ func ToBytes4(input []byte) (B4, error) {
 		)
 	}
 	return B4(input), nil
+}
+
+// FromUint32 returns a new B4 from a uint32.
+func FromUint32(v uint32) B4 {
+	h := B4{}
+	binary.LittleEndian.PutUint32(h[:], v)
+	return h
+}
+
+// ToUint32 returns a new uint32 from a B4.
+func (h B4) ToUint32() uint32 {
+	return binary.LittleEndian.Uint32(h[:])
 }
 
 /* -------------------------------------------------------------------------- */
@@ -91,3 +104,7 @@ func (h B4) MarshalSSZ() ([]byte, error) {
 func (h B4) HashTreeRoot() (B32, error) {
 	return ToBytes32(ExtendToSize(h[:], B32Size))
 }
+
+/* -------------------------------------------------------------------------- */
+/*                            uint32 conversion                               */
+/* -------------------------------------------------------------------------- */
